fix ci unittest warnings
