+++ sdk/python/feast/on_demand_pandas_transformation.py
@@ -30,9 +30,6 @@ class OnDemandPandasTransformation:
                 "Comparisons should only involve OnDemandPandasTransformation class objects."
             )
 
-        if not super().__eq__(other):
-            return False
-
         if (
             self.udf_string != other.udf_string
             or self.udf.__code__.co_code != other.udf.__code__.co_code
