+++ sdk/python/tests/utils/test_wrappers.py
@@ -1,14 +1,14 @@
-import pytest
+import warnings
 
 
 def no_warnings(func):
     def wrapper_no_warnings(*args, **kwargs):
-        with pytest.warns(None) as warnings:
+        with warnings.catch_warnings(record=True) as record:
             func(*args, **kwargs)
 
-        if len(warnings) > 0:
+        if len(record) > 0:
             raise AssertionError(
-                "Warnings were raised: " + ", ".join([str(w) for w in warnings])
+                "Warnings were raised: " + ", ".join([str(w) for w in record])
             )
 
     return wrapper_no_warnings
